+++ src/RecoveryServices/RecoveryServices/Models/PSObjects.cs
@@ -82,6 +82,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices
             this.Properties.ProvisioningState = vault.Properties.ProvisioningState;
             this.Properties.PrivateEndpointStateForBackup = vault.Properties.PrivateEndpointStateForBackup;
             this.Properties.PrivateEndpointStateForSiteRecovery = vault.Properties.PrivateEndpointStateForSiteRecovery;
+            this.Identity = vault.Identity;
         }
 
         #endregion
@@ -122,6 +123,11 @@ namespace Microsoft.Azure.Commands.RecoveryServices
         /// </summary>
         public ARSVaultProperties Properties { get; set; }
 
+        /// <summary>
+        /// Gets or sets Identity.
+        /// </summary>
+        public IdentityData Identity { get; set; }
+
         #endregion
     }
 
@@ -159,6 +165,11 @@ namespace Microsoft.Azure.Commands.RecoveryServices
         /// </summary>
         public string BackupStorageRedundancy { get; set; }
 
+        /// <summary>
+        /// Gets or sets CrossRegionRestore Flag.
+        /// </summary>
+        public bool CrossRegionRestore { get; set; }
+
         #endregion
     }
 
@@ -194,7 +205,17 @@ namespace Microsoft.Azure.Commands.RecoveryServices
     public enum AzureRmRecoveryServicesBackupStorageRedundancyType
     {
         GeoRedundant = 1,
+        ZoneRedundant,
         LocallyRedundant
+    }
+
+    /// <summary>
+    /// Class to define the vault BackupStorageRedundancy settings.
+    /// </summary>
+    public enum MSIdentity
+    {
+        SystemAssigned = 1,
+        None
     }
 
     /// <summary>