+++ src/Synapse/Synapse/Models/PSDataFlowResource.cs
@@ -1,5 +1,4 @@
 ﻿using Azure.Analytics.Synapse.Artifacts.Models;
-using Newtonsoft.Json;
 
 namespace Microsoft.Azure.Commands.Synapse.Models
 {
@@ -18,13 +17,5 @@ namespace Microsoft.Azure.Commands.Synapse.Models
         public string WorkspaceName { get; set; }
 
         public DataFlow Properties { get; set; }
-
-        [JsonProperty(PropertyName = "properties")]
-        internal PSDataFlow PropertiesForCreate { get; set; }
-
-        public DataFlowResource ToSdkObject()
-        {
-            return new DataFlowResource(this.PropertiesForCreate?.ToSdkObject());
-        }
     }
 }