+++ src/Synapse/Synapse/Models/SynapseConstants.cs
@@ -32,6 +32,8 @@ namespace Microsoft.Azure.Commands.Synapse.Models
 
         public const string SqlPool = nameof(SqlPool);
 
+        public const string SqlPoolName = nameof(SqlPoolName);
+
         public const string RestorePoint = nameof(RestorePoint);
 
         public const string Sql = nameof(Sql);
@@ -130,6 +132,12 @@ namespace Microsoft.Azure.Commands.Synapse.Models
         public const string Dataset = nameof(Dataset);
 
         public const string DataFlow = nameof(DataFlow);
+
+        public const string SensitivityRecommendation = nameof(SensitivityRecommendation);
+
+        public const string PointInTime = nameof(PointInTime);
+
+        public const string TargetSqlPoolName = nameof(TargetSqlPoolName);
 
         public static Dictionary<string, ComputeNodeSize> ComputeNodeSizes = new Dictionary<string, ComputeNodeSize>
         {