+++ tools/Az.Tools.Predictor/Az.Tools.Predictor.Test/Mocks/MockAzPredictorService.cs
@@ -13,6 +13,7 @@
 // ----------------------------------------------------------------------------------
 
 using System.Collections.Generic;
+using System.Management.Automation.Language;
 
 namespace Microsoft.Azure.PowerShell.Tools.AzPredictor.Test.Mocks
 {
@@ -22,9 +23,14 @@ namespace Microsoft.Azure.PowerShell.Tools.AzPredictor.Test.Mocks
     sealed class MockAzPredictorService : AzPredictorService
     {
         /// <summary>
-        /// Gets or sets if a predictions is requested.
+        /// Gets or sets the commands in history to request prediction for.
         /// </summary>
-        public bool IsPredictionRequested { get; set; }
+        public IEnumerable<string> Commands { get; set; }
+
+        /// <summary>
+        /// Gets or sets the commands that's recorded in history.
+        /// </summary>
+        public CommandAst History { get; set; }
 
         /// <summary>
         /// Constructs a new instance of <see cref="MockAzPredictorService"/>
@@ -32,23 +38,40 @@ namespace Microsoft.Azure.PowerShell.Tools.AzPredictor.Test.Mocks
         /// <param name="history">The history that the suggestion is for</param>
         /// <param name="suggestions">The suggestions collection</param>
         /// <param name="commands">The commands collection</param>
-        public MockAzPredictorService(string history, IList<string> suggestions, IList<string> commands)
+        public MockAzPredictorService(string history, IList<PredictiveCommand> suggestions, IList<PredictiveCommand> commands)
         {
-            SetPredictionCommand(history);
-            SetCommandsPredictor(commands);
-            SetSuggestionPredictor(history, suggestions);
+            if (history != null)
+            {
+                SetCommandToRequestPrediction(history);
+
+                if (suggestions != null)
+                {
+                    SetCommandBasedPreditor(history, suggestions);
+                }
+            }
+
+            if (commands != null)
+            {
+                SetFallbackPredictor(commands);
+            }
         }
 
         /// <inheritdoc/>
-        public override void RequestPredictions(IEnumerable<string> history)
+        public override void RequestPredictions(IEnumerable<string> commands)
         {
-            this.IsPredictionRequested = true;
+            Commands = commands;
         }
 
         /// <inheritdoc/>
-        protected override void RequestCommands()
+        protected override void RequestAllPredictiveCommands()
         {
             // Do nothing since we've set the command and suggestion predictors.
+        }
+
+        /// <inheritdoc/>
+        public override void RecordHistory(CommandAst history)
+        {
+            History = history;
         }
     }
 }